requirements.txt
streamlit
pytesseract
pillow
PyQt5